add make file
1.add make file
2.build fast/super power by different cpp file
