+++ testfastpower.cpp
@@ -2,6 +2,7 @@
 #include <sstream>
 #include <string>
 #include <vector>
+#include "fast_power.h"
 
 using namespace std;
 
@@ -16,9 +17,7 @@ using namespace std;
         stringstream(in) >> out;\
     }while(0)
 
-#if FAST_POWER
 
-#include "fast_power.h"
 int main(int argc, char **argv)
 {
     long long base = 0;
@@ -30,22 +29,3 @@ int main(int argc, char **argv)
     cout << fast_power(base, power , mod) << endl;
     return 0;
 }
-#else
-
-#include "super_power.h"
-
-int main(int argc, char **argv)
-{
-    long long base = 0;
-    unsigned long long mod = -1ull;
-    string power_str = argv[2];
-    vector<unsigned char> power_vec(power_str.size());
-    chect_param_count(3);
-    ton(argv[1], base);
-    if(argc > 3) ton(argv[3], mod);
-    for(unsigned long long i = 0; i < power_str.size(); i++)
-        power_vec[i] = power_str[i] - '0';
-    cout << super_power(base, power_vec.data(), power_vec.size(), mod) << endl;
-    return 0;
-}
-#endif
